+++ Module.php
@@ -50,11 +50,6 @@ class Module extends ContentContainerModule
     /**
      * @inheritdoc
      */
-    public $resourcesPath = 'resources';
-
-    /**
-     * @inheritdoc
-     */
     public function init()
     {
         parent::init();